fix: ansible versions validation tasks
fix #840

Since a6712e391ab1a283cf24102a4055ea16fc756105, ansible and jinja versions are
validated to ensure supported versions are used.
This validation is done by delegating `command` tasks to localhost, and parsing
the standard output of the executed commands, e.g. `ansible --version`.

The `ansible.builtin.command` module is not run when in check mode, causing the
variable which is supposed to get the result of the command to be empty,
resulting in an error message in the following tasks that parse these result
variables.

This commit ensures the `command` tasks are run even in check_mode. As they do
no modification on localhost, this is not dangerous.

Signed-off-by: Alexis Thietard <alexis.thietard@rtone.fr>